+++ hooks/lando-setup-build-engine-darwin.js
@@ -1,5 +1,6 @@
 'use strict';
 
+const axios = require('../utils/get-axios')();
 const os = require('os');
 const path = require('path');
 const semver = require('semver');
@@ -86,6 +87,9 @@ module.exports = async (lando, options) => {
   // cosmetics
   const install = version ? `v${version}` : `build ${build}`;
 
+  // download url
+  const url = getEngineDownloadUrl(build);
+
   // darwin install task
   options.tasks.push({
     title: `Downloading build engine`,
@@ -112,8 +116,8 @@ module.exports = async (lando, options) => {
     canRun: async () => {
       // throw if we cannot resolve a semantic version to a buildid
       if (!build) throw new Error(`Could not resolve ${install} to an installable version!`);
-      // throw error if not online
-      if (!await require('is-online')()) throw new Error('Cannot detect connection to internet!');
+      // throw error if we cannot ping the download link
+      await axios.head(url);
       // throw if user is not an admin
       if (!await require('../utils/is-admin-user')()) {
         throw new Error([
@@ -127,7 +131,7 @@ module.exports = async (lando, options) => {
     task: async (ctx, task) => {
       try {
         // download the installer
-        ctx.download = await downloadDockerDesktop(getEngineDownloadUrl(build), {ctx, debug, task});
+        ctx.download = await downloadDockerDesktop(url, {ctx, debug, task});
 
         // prompt for password if interactive
         if (ctx.password === undefined && lando.config.isInteractive) {